pyproject.toml
[project]
name = "searchconsole-mcp"
version = "0.1.0"
description = "MCP server for Google Search Console — query search analytics, manage sitemaps, inspect URLs"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
authors = [
    {name = "Haris Nadeem", email = "haris@harisnadeem.com"},
]
keywords = ["mcp", "google", "search-console", "seo", "analytics", "claude", "cursor", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "google-auth>=2.40.0",
    "httpx>=0.28.1",
    "mcp>=1.24.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0",
    "ruff>=0.4.0",
    "pytest>=8.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
]

[project.urls]
Homepage = "https://github.com/harisnadeem/searchconsole-mcp"
Repository = "https://github.com/harisnadeem/searchconsole-mcp"
Issues = "https://github.com/harisnadeem/searchconsole-mcp/issues"
Changelog = "https://github.com/harisnadeem/searchconsole-mcp/blob/main/CHANGELOG.md"

[project.scripts]
searchconsole-mcp = "searchconsole_mcp.server:run_server"

[build-system]
requires = ["setuptools>=82.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["searchconsole_mcp*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "D", "UP", "B", "C4", "SIM"]
ignore = ["D203", "D212"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]
